+++ BlinkID/src/android/java/com/phonegap/plugins/microblink/overlays/serialization/BarcodeOverlaySettingsSerialization.java
@@ -1,5 +1,6 @@
 package com.phonegap.plugins.microblink.overlays.serialization;
 
+import com.microblink.blinkid.R;
 import com.microblink.entities.recognizers.RecognizerBundle;
 import com.microblink.uisettings.BarcodeUISettings;
 import com.microblink.uisettings.UISettings;
@@ -11,7 +12,9 @@ public final class BarcodeOverlaySettingsSerialization implements OverlaySetting
     @Override
     public UISettings createUISettings(JSONObject jsonUISettings, RecognizerBundle recognizerBundle) {
         // no settings deserialized at the moment
-        return new BarcodeUISettings(recognizerBundle);
+        BarcodeUISettings uiSettings = new BarcodeUISettings(recognizerBundle);
+        uiSettings.setSplashScreenLayoutResourceID(R.layout.mb_layout_splash_screen);
+        return uiSettings;
     }
 
     @Override