+++ Release notes.md
@@ -1,6 +1,7 @@
 ## 4.2.0
 - Updated to [Android SDK v4.2.0](https://github.com/BlinkID/blinkid-android/releases/tag/v4.2.0) and [iOS SDK v4.2.0](https://github.com/BlinkID/blinkid-ios/releases/tag/v4.2.0)
 - [android] fixed NPE when serializing recognizer result containing `null` byte array field 
+- [android] removed Microblink logo from camera splash screen
 
 
 ## 4.1.0